+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## v0.4.0 2024-04-20
+
+- Add Cedar human-readable schema support
+- Update copyright to Cedar Contributors
+
 ## v0.3.0 2024-01-20
 
 - Update pattern matching